Allow right click for meta blocks
When right-clicking directly on selectable text, the browser will select the current hovered word before invoking context menu callback. And our callback detects if any text is selected and aborts custom context menu if so.

This allows right-click on meta-block text by disabled selection of meta block content until item has been selected first. This makes meta-blocks work like the main text block for the purpose of focus-before-selection as well as showing context menu on first click when unselected, but not when selected.

This effort is only fully implemented for title meta block in this commit, but can easily be extended to the other meta blocks.