Add the -o flag to composer install for production build processes
This flag is recommended for production only:
"Note: You should not enable any of these optimizations in development as they all will cause various problems when adding/removing classes. The performance gains are not worth the trouble in a development setting."

@see https://getcomposer.org/doc/articles/autoloader-optimization.md#autoloader-optimization
Maybe this should be added to .travis.yml
But maybe it could cause issues with the
cached builds.
It looks like the performance improvement could be
very slight, compared to possibly debugging
in the future.