Added initial Focusable functionality
Currently it's only in use for the tabs in the Library. If you focus any tab in the library, pressing tab and shift-tab will move correctly back and forwards between the tabs and pressing enter will switch selection to the focused tab.

However, pressing left and right arrows will also navigate to the other tabs (even observing reading direction), but of course not outside the tab list.

This should be fully accessible and semantic..