Amp-Wp-App-Shell - Improve flexibility for different types of caching
Update amp-wp-app-shell.js

**Rationale**
Due to caching restrictions with WPengine the use of a query-string parameter to
denote an inner shell request is problematic as it prevents the request from
being cached independently. The request needs to be distinctly different from
the outer shell in-order for it to be treated as a seperate cache item.

**Changes**

- Query string parameter is now appended to the path instead of
- Fixed bug where code was always expecting a .site-branding element to exist