Use new story reducer in stead of old states
This commit actually puts the new story reducer in use and it's implemented everywhere story state is changed.

It however still has one weird problem - when a new element is selected, it somehow sticks to the mouse. I have debugged it much, hoping that @miina knows how to fix it.