Improve security of external href (make it harder for malicious people).
Initially I reported this via the Google Seucrity program (in
https://issuetracker.google.com/issues/151134576), but it was
rejected with this explanation:
https://sites.google.com/site/bughunteruniversity/nonvuln/phishing-with-window-opener

While the change may not fix all possible attack vectors, it at least
takes care about one (window.opener). It add one more layer to the
security onion, and makes it one step harder for malicious people.

The probability that this issues can be exploitet is low, as amp-wp.org
needs to be compromised, and an admin of an wordpress installation needs
to click on the getting started link.

The security impact is medium, as a breach can theoretically affect a
lot of websites (currently 400.000+ installations counted on the
wordpress plugin page for this plugin) and there are other ways to
do malicious things which are not prevented by this change.

There are also static security scanners (commercial and open source)
which flag this as a blocking security issue. Adding this change will
make that this is not showing up as an issue in such scans, which may
make a difference for companies where such scans are mandatory to not
show up security issues.

All in all this is a security related improvement, but not a big security
hole.