Allow filter of the inner component URL
In some cases, the plugin integrator may need to alter the format of the
inner component URL.

While it is possible to change the way incoming requests for inner
components are handled on the back-end (PHP), there was no way to change
the URL format of the request on the front-end (JS).

This commit makes the inner component request URL filterable using
wp.hooks. Here's how you could filter the URL in your plugin or theme
so that instead of using a query parameter, a new segment is added to
the URL:

```
if ( ampAppShell && ampAppShell.hooks ) {
	function filterInnerComponentUrl( url, baseUrl, componentQueryVar ) {
		const filteredUrl = new URL( baseUrl );
		const pathSuffix = '_' + componentQueryVar + '_inner';
		filteredUrl.pathname = filteredUrl.pathname + pathSuffix;

		return filteredUrl;
	}

	ampAppShell.hooks.addFilter( 'amp.appShell.innerComponentUrl', 'ampWpAppShell/filterInnerComponentUrl', filterInnerComponentUrl );
}
```