Use custom hook for determining post dirtiness state
The WordPress core `isDirty()` selector gets cleared only if a post gets published or updated. In case of an autosave, it still returns `true`.

In order to allow the users to revalidate AMP only if something changed since the last autosave, we had to add a custom hook that tracks the post content changes.

Once the first update is detected, a flag `isPostDirty` flag is set in the store and content is no longer checked (it's expensive since each time a content is requested, the `blocks` list must be serialized).

When the post gets saved or autosaved, the flag is cleared and the content is observed again.