Handle CSS build process with Webpack
So far, all stylesheets that were not imported in JS files were handled with plain CLI commands. The simple "build" process consisted of copying all CSS files from `./assets/css/src` to `./assets/css` and generating RTL stylesheets with the `rtlcss` CLI command.

With this update, all files in the `./assets/css/src` folder are handled by the new 'Styles' Webpack entry point. Thanks to that, they are watched and rebuilt on-the-fly with the regular `npm run dev` task. Also, it is now possible to use the SASS/SCSS preprocessing, if needed.