Handle case when no scannable URL is available
In the Reader mode, only post types are used for performing a site scan. If there are no posts, pages or other custom post types available for a scan, a request made to the `/wp-json/amp/v1/scannable-urls` endpoint returns an empty array.

With this change, an empty array is treated as a valid response so that the list of scannable URLs kept in the Site Scan state is emptied. This case is then handled gracefully on the frontend where an error message is rendered. The message says that there are no URLs available for a scan.

Even though the error message has already been part of the Site Scan UI, it was displayed only if there were no scannable URLs available initially, right after a page load.