include $parsed_url[path] in the assessment
There are instances where a wordpress install is implemented as a backend but served as though in a subdirectory path. 
i.e. backend is https://some_wordpress_install/ but served to the user as https://example.com/blog/. This fix makes sure to add the path of the home_url and removes that part from the $_SERVER['REQUEST_URI']. This way, the right uri is always preserved, otherwise it breaks.